+++ velesresearch/wrappers.py
@@ -62,6 +62,7 @@ def survey(
     questionTitlePattern: str = "numTitleRequire",
     requiredText: str = "*",
     scoresSuffix: str = "_score",
+    showCompleteButton: bool = True,
     showCompletedPage: bool = True,
     showNavigationButtons: str = "bottom",
     showPageNumbers: bool | None = None,
@@ -142,6 +143,7 @@ def survey(
         questionTitlePattern (str): The pattern of the question title. See <https://surveyjs.io/form-library/documentation/design-survey/configure-question-titles#title-pattern>.
         requiredText (str): The text denoting the required questions. Default is '*'.
         scoresSuffix (str): The suffix of the score column if `addScoreToResults=True`. Default is '_score'.
+        showCompleteButton (bool): Whether to show the 'Complete' button. Default is True.
         showCompletedPage (bool): Whether to show the completed page. Default is True.
         showNavigationButtons (str): The location of the navigation buttons. Can be 'bottom' (default), 'top', 'both', 'none'.
         showPageNumbers (bool | None): Whether to show the page numbers in the pages' titles.
@@ -223,6 +225,7 @@ def survey(
         "questionTitlePattern": questionTitlePattern,
         "requiredText": requiredText,
         "scoresSuffix": scoresSuffix,
+        "showCompleteButton": showCompleteButton,
         "showCompletedPage": showCompletedPage,
         "showNavigationButtons": showNavigationButtons,
         "showPageNumbers": showPageNumbers,